+++ packages/dashboard/src/app/views/myStories/index.js
@@ -17,7 +17,7 @@
 /**
  * External dependencies
  */
-import { useEffect, useMemo } from '@web-stories-wp/react';
+import { useEffect, useMemo, useCallback } from '@web-stories-wp/react';
 
 /**
  * Internal dependencies
@@ -41,10 +41,12 @@ function MyStories() {
     storiesOrderById,
     totalPages,
     totalStoriesByStatus,
+    getAuthors,
   } = useApi(
     ({
       actions: {
         storyApi: { duplicateStory, fetchStories, trashStory, updateStory },
+        usersApi: { getAuthors },
       },
       state: {
         stories: {
@@ -67,6 +69,7 @@ function MyStories() {
       storiesOrderById,
       totalPages,
       totalStoriesByStatus,
+      getAuthors,
     })
   );
 
@@ -76,6 +79,30 @@ function MyStories() {
       isLoading,
       totalPages,
     });
+
+  const { setQueriedAuthors } = author;
+  const queryAuthorsBySearch = useCallback(
+    (authorSearchTerm) => {
+      return getAuthors(authorSearchTerm).then((data) => {
+        const userData = data.map(({ id, name }) => ({
+          id,
+          name,
+        }));
+        setQueriedAuthors((exisitingUsers) => {
+          const exisitingUsersIds = exisitingUsers.map(({ id }) => id);
+          const newUsers = userData.filter(
+            (newUser) => !exisitingUsersIds.includes(newUser.id)
+          );
+          return [...exisitingUsers, ...newUsers];
+        });
+      });
+    },
+    [getAuthors, setQueriedAuthors]
+  );
+
+  useEffect(() => {
+    queryAuthorsBySearch();
+  }, [queryAuthorsBySearch]);
 
   useEffect(() => {
     fetchStories({
@@ -113,6 +140,7 @@ function MyStories() {
         totalStoriesByStatus={totalStoriesByStatus}
         view={view}
         author={author}
+        queryAuthorsBySearch={queryAuthorsBySearch}
       />
 
       <Content
